chore: remove scripts_regression_tests.py in favor of pytest
